- Add cross-server notification (using BungeeCord Message channel) - Add ability to read any mail by ID (perm: SimpleMail.spy) - Anyone can read their own outbox mail (via /readmail <id>) - Make dates a little more readable - Clean up A LOT of bad code - Made some of the SQL more efficient (like the purge!) - Fix several issues where ResultSet was not initialised before use

+++ src/me/odium/test/commands/delmail.java
@@ -39,13 +39,14 @@ public class delmail implements CommandExecutor {
       stmt = con.createStatement();
 
       rs = stmt.executeQuery("SELECT * FROM SM_Mail WHERE id='" + args[0] + "'");
-
-      if (!rs.getString("target").equalsIgnoreCase(Playername)) {
-        sender.sendMessage(plugin.GRAY+"[SimpleMail] "+plugin.RED+"This is not your message to delete or it does not exist. ");
-      } else {
-        stmt.executeUpdate("DELETE FROM SM_Mail WHERE id='"+args[0]+"' AND target='"+Playername+"'");
-        sender.sendMessage(plugin.GRAY+"[SimpleMail] "+plugin.GREEN+"Message Deleted.");
-      } 
+      if (rs.next()) {
+        if (!rs.getString("target").equalsIgnoreCase(Playername)) {
+          sender.sendMessage(plugin.GRAY+"[SimpleMail] "+plugin.RED+"This is not your message to delete or it does not exist. ");
+        } else {
+          stmt.executeUpdate("DELETE FROM SM_Mail WHERE id='"+args[0]+"' AND target='"+Playername+"'");
+          sender.sendMessage(plugin.GRAY+"[SimpleMail] "+plugin.GREEN+"Message Deleted.");
+        }
+      }
       rs.close();
     } catch(Exception e) {         
       if (e.toString().contains("ResultSet closed")) {
